refactor: drop deadcode
The functions `networkHeartbeat()` and `checkDownloadSpeed()` take an
array of URLs as argument. This array doesn't exist until after the
Trader is started and it gets a good poll reply from the server.
Therefore, it is useless to run this at this point, they'll do nothing.
Additionally, `initNetworkMeasurements()` is run right after the current
function, which starts the ping and speedtest intervals with the
necessary arguments.